[Guardian] Organize shared type modules
