+++ crates/hashi-guardian/src/s3_reader/heartbeat_checks.rs
@@ -6,10 +6,10 @@ use super::VerifiedLogRecord;
 use crate::HEARTBEAT_INTERVAL;
 use crate::LIVE_SESSION_LATEST_HEARTBEAT_MAX_AGE;
 use crate::OTHER_SESSION_QUIET_PERIOD;
-use hashi_types::guardian::s3_utils::S3HourScopedDirectory;
-use hashi_types::guardian::time_utils::now_timestamp_secs;
-use hashi_types::guardian::time_utils::unix_millis_to_seconds;
-use hashi_types::guardian::time_utils::UnixSeconds;
+use hashi_types::guardian::s3::S3HourScopedDirectory;
+use hashi_types::guardian::time::now_timestamp_secs;
+use hashi_types::guardian::time::unix_millis_to_seconds;
+use hashi_types::guardian::time::UnixSeconds;
 use hashi_types::guardian::GuardianError::CurrentSessionHeartbeatNotLive;
 use hashi_types::guardian::GuardianError::InvalidS3Log;
 use hashi_types::guardian::GuardianError::PriorSessionHeartbeatStillRecent;
