+++ crates/hashi-types/src/guardian/s3/log_messages/kp_share.rs
@@ -1,8 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use super::super::ObjectKeyPattern;
-use super::super::S3_DIR_KP_SHARES;
+use super::super::log_layout::ObjectKeyPattern;
+use super::super::log_layout::S3_DIR_KP_SHARES;
 use crate::guardian::GuardianError;
 use crate::guardian::KPEncryptedShares;
 use crate::guardian::KPEncryptedSharesRoster;
