+++ crates/hashi-types/src/guardian/s3/mod.rs
@@ -1,0 +1,12 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+mod config;
+pub mod log;
+mod log_layout;
+mod log_messages;
+mod log_record;
+mod log_schema;
+
+pub use config::*;
+pub use log_layout::S3HourScopedDirectory;
